Move featureFlagDecider to featureFlag folder
